HT_24/requirements.txt
factory_boy==3.3.0
django_filter==23.5
Markdown==3.5.2
djangorestframework==3.14.0
redis==5.0.1
requests==2.31.0
celery==5.3.6
python_dotenv==1.0.1
django==5.0.1